Refactor R serialization.
