+++ src/programming/Ligare/programming/R/__init__.py
@@ -1,9 +1,10 @@
 from Ligare.programming.R.exception import RscriptProcessError, RscriptScriptError
 from Ligare.programming.R.process import RProcessStepBuilder
-from Ligare.programming.R.type_conversion import (
+from Ligare.programming.R.serializers import (
     boolean,
-    convert,
+    composite_type_from_parts,
     from_parts,
+    number,
     number_from_csv,
     number_from_parts,
     number_vector_from_csv,
@@ -11,7 +12,6 @@ from Ligare.programming.R.type_conversion import (
     string_from_csv,
     string_from_parts,
     string_vector_from_csv,
-    vector_from_csv,
 )
 
 __all__ = (
@@ -19,8 +19,9 @@ __all__ = (
     "RscriptProcessError",
     "RscriptScriptError",
     "boolean",
-    "convert",
+    "composite_type_from_parts",
     "from_parts",
+    "number",
     "number_from_csv",
     "number_from_parts",
     "number_vector_from_csv",
@@ -28,5 +29,4 @@ __all__ = (
     "string_from_csv",
     "string_from_parts",
     "string_vector_from_csv",
-    "vector_from_csv",
 )
